+++ .gitignore
@@ -11,5 +11,6 @@ config.ini
 temp/*
 
 .venv
+.vscode
 
 __pycache__
