+++ README.md
@@ -46,4 +46,6 @@ flake8
 
 4. Вынести UI в отдельный слой приложения, отделить от логики модели, в сторону Layered или Hexagonal архитектуры
 
+5. Сохранять картинку, аудио и видео по случайному пути без перезаписи предыдущего результата, чтобы добавить возможность кэширования, параллельной работы для нескольких пользователей (например для возможного веб интерфейса)
+
 5. Добавить тесты, запускать в github actions
