+++ domain/speech_synt.py
@@ -1,12 +1,16 @@
 import os
-
+import uuid
 import torch
 
 
-def get_speech(text, selected_speaker):
+def get_speech(text, selected_speaker) -> str:
+    """
+    return audio file path
+    """
     device = torch.device("cpu")
     torch.set_num_threads(4)
-    local_file = "model.pt"
+
+    local_file = 'model.pt'
 
     if not os.path.isfile(local_file):
         torch.hub.download_url_to_file(
@@ -18,10 +22,14 @@ def get_speech(text, selected_speaker):
 
     sample_rate = 48000
 
+    audio_path = 'data/'+str(uuid.uuid4())+'.wav'
+
     audio_paths = model.save_wav(
-        text=text, speaker=selected_speaker, sample_rate=sample_rate
+        text=text, speaker=selected_speaker, sample_rate=sample_rate, audio_path=audio_path
     )
-    print(audio_paths)
+    # print(audio_paths)
+
+    return audio_path
 
 
-get_speech("text", "en_99")
+# get_speech("text", "en_99")
