+++ native_app.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+
+from ui.native_app.app import run
+
+
+if __name__ == '__main__':
+    run()
