+++ ui/native_app/clickable_slider.py
